+++ pyArango/document.py
@@ -193,6 +193,7 @@ class Document(object) :
         self.connection = self.collection.connection
         self.documentsURL = self.collection.documentsURL
 
+
         self.URL = None
         self.setPrivates(jsonFieldInit)
         self._store = DocumentStore(self.collection, validators=self.collection._fields, initDct=jsonFieldInit)
@@ -257,7 +258,7 @@ class Document(object) :
                 if update :
                     self._rev = data['_rev']
                 else :
-                    self.setPrivates(data)
+                    self.set(data)
             else :
                 if update :
                     raise UpdateError(data['errorMessage'], data)
@@ -401,7 +402,7 @@ class Edge(Document) :
 
         if fromVertice.__class__ is Document :
             if not fromVertice._id :
-                fromVertice._id.save()
+                fromVertice.save()
 
             self._from = fromVertice._id
         elif (type(fromVertice) is bytes) or (type(fromVertice) is str) :
@@ -409,7 +410,7 @@ class Edge(Document) :
 
         if toVertice.__class__ is Document :
             if not toVertice._id :
-                toVertice._id.save()
+                toVertice.save()
 
             self._to = toVertice._id
         elif (type(toVertice) is bytes) or (type(toVertice) is str) :