+++ CHANGELOG.rst
@@ -22,6 +22,12 @@
 * REFACT: New DocumentStore class for taking care of storing document in a hierarchy of stores (nested objects) and validate them
 * Minor bug fixes
 
+1.2.9
+=====
+
+* Added bulk import to connection
+* Added binvars to explain
+
 1.2.8
 =====